working on transformation

+++ framework/camera.hpp
@@ -9,7 +9,7 @@ struct Camera{
     glm::vec3 pos_{0.0f,0.0f,0.0f};
     glm::vec3 direction_{0.0f,0.0f,-1.0f};
     glm::vec3 up_{0.0f,1.0f,0.0f};
-    unsigned int fovX{45};
+    float fovX{45};
 
 };
